Small fixes/tidy-ups
Added () to split method called on author.

Tidied up the for-loop that obtains document views. If there doesn't exist a next item in the iterator, we catch the exception raises by __next__ and then break from the loop.

Coupled the if statement that checked if depth was >= 5 with the one that checked if we'd already traversed the author. Moved them to an appropriate place to avoid unnecessary work being done if we should be returning anyway.